backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]==0.24.0
tortoise-orm[psycopg]==0.25.1
tortoise-vector==0.1.4
asyncpg==0.29.0
python-dotenv==1.0.0
numpy==1.25.2
pydantic==2.5.0
python-multipart==0.0.6
aerich==0.9.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
email-validator>=2.2.0